ui consistency updated/added to whole design
