+++ tests/test_uix.py
@@ -5,7 +5,7 @@ This file is responsible for testing custom UIX from designer/uix/*
 import unittest
 from nose.tools import assert_equal
 from nose.tools import assert_not_equal
-from designer.uix.actioncheckbutton import ActionCheckButton
+from designer.uix.designer_action_items import ActionCheckButton
 from designer.uix.kivy_console import KivyConsole
 from designer.uix.settings import SettingListCheckItem
